requirements.txt
# Core framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy>=2.0.36
alembic>=1.14.0

# Auth
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0

# Forms & templates
python-multipart>=0.0.12
jinja2>=3.1.4
itsdangerous>=2.2.0

# Config
python-dotenv>=1.0.1

# Validation
pydantic[email]>=2.10.0

# Testing
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Optional: PostgreSQL support (uncomment for production)
# psycopg2-binary==2.9.9
